utils/token_cleaner/requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.8.0
python-dotenv==1.1.1
Werkzeug==3.0.1
pymongo==4.6.0
dnspython==2.4.2
setuptools
gunicorn==21.2.0
requests==2.31.0
flask-cors==5.0.0
jwt-auth-middleware @ git+https://github.com/Hsieh-Yu-Hung/JWT_Midware.git